hearth_bot/requirements.txt
aiogram>=3.0.0
aiohttp
python-dotenv
aiosqlite
google-generativeai
duckduckgo-search
chromadb
sentence-transformers
